+++ pmd-core/src/main/java/net/sourceforge/pmd/reporting/ParametricRuleViolation.java
@@ -7,17 +7,14 @@ package net.sourceforge.pmd.reporting;
 import java.util.Collections;
 import java.util.Map;
 
-import net.sourceforge.pmd.annotation.InternalApi;
 import net.sourceforge.pmd.lang.document.FileLocation;
 import net.sourceforge.pmd.lang.rule.Rule;
 import net.sourceforge.pmd.util.AssertionUtil;
 
 /**
- * @deprecated This is internal. Clients should exclusively use {@link RuleViolation}.
+ * @apiNote This is internal API. Clients should exclusively use {@link RuleViolation}.
  */
-@Deprecated
-@InternalApi
-public class ParametricRuleViolation implements RuleViolation {
+class ParametricRuleViolation implements RuleViolation {
     protected final Rule rule;
     protected final String description;
 
@@ -25,25 +22,11 @@ public class ParametricRuleViolation implements RuleViolation {
 
     private final Map<String, String> additionalInfo;
 
-    // todo add factory methods on the interface and hide the class.
-
-    /**
-     * @deprecated Update tests that use this not to call the ctor directly.
-     */
-    @Deprecated
-    public ParametricRuleViolation(Rule theRule, Reportable node, String message) {
+    ParametricRuleViolation(Rule theRule, Reportable node, String message) {
         this(theRule, node.getReportLocation(), message, Collections.emptyMap());
     }
 
-    public ParametricRuleViolation(Rule theRule, FileLocation location, String message) {
-        this(theRule, location, message, Collections.emptyMap());
-    }
-
-    public ParametricRuleViolation(Rule theRule, Reportable node, String message, Map<String, String> additionalInfo) {
-        this(theRule, node.getReportLocation(), message, additionalInfo);
-    }
-
-    public ParametricRuleViolation(Rule theRule, FileLocation location, String message, Map<String, String> additionalInfo) {
+    ParametricRuleViolation(Rule theRule, FileLocation location, String message, Map<String, String> additionalInfo) {
         this.rule = AssertionUtil.requireParamNotNull("rule", theRule);
         this.description = AssertionUtil.requireParamNotNull("message", message);
         this.location = location;