+++ pmd-core/src/test/java/net/sourceforge/pmd/renderers/XSLTRendererTest.java
@@ -12,7 +12,6 @@ import net.sourceforge.pmd.FooRule;
 import net.sourceforge.pmd.lang.document.FileId;
 import net.sourceforge.pmd.lang.document.FileLocation;
 import net.sourceforge.pmd.lang.document.TextRange2d;
-import net.sourceforge.pmd.reporting.ParametricRuleViolation;
 import net.sourceforge.pmd.reporting.Report;
 import net.sourceforge.pmd.reporting.RuleViolation;
 
@@ -59,7 +58,7 @@ class XSLTRendererTest extends AbstractRendererTest {
     void testDefaultStylesheet() throws Exception {
         XSLTRenderer renderer = new XSLTRenderer();
         FileLocation loc = FileLocation.range(FileId.UNKNOWN, TextRange2d.range2d(1, 1, 1, 2));
-        RuleViolation rv = new ParametricRuleViolation(new FooRule(), loc, "violation message");
+        RuleViolation rv = newRuleViolation(new FooRule(), loc, "violation message");
         String result = renderReport(renderer, it -> it.onRuleViolation(rv));
         assertTrue(result.contains("violation message"));
     }